fix: move backup and restore functions to new class
The Pool class can not be sent over a port. The backup and restore
functions were moved out of the entries database to remove this
dependency.
